+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Doesn't overwrite `Content-Type` if already set.
+- Renames to FetchJsonify.
 
 ## 0.1.338 (Dec 12, 2015)
 - Deletes the `json` property before calling `fetch`.